add OVERRIDE to ChiSquareMetric
message describing the commit

+++ modules/statistics/src/ChiSquareMetric.cpp
@@ -84,7 +84,7 @@ double ChiSquareMetric::get_distance(unsigned i, unsigned j) const
  }
 }
 
-Float ChiSquareMetric::get_weight(unsigned i) const
+double ChiSquareMetric::get_weight(unsigned i) const
 {
  return weight_[i];
 }
